SQL D.A.X on Customer Behavior
What are some of the reasons a typical e-commerce customer will browse the site but not buy until a later visit?

There is no popular answer, however, one popular reason is comparison shopping between different e-commerce sites before ultimately making a purchase decision. This is very common for luxury goods where significant up-front research and comparison is required by the customer before deciding (think car purchases) but also true to a lesser extent for the merchandise on this site (t-shirts, accessories, etc).
In the world of online marketing, identifying and marketing to these future customers based on the characteristics of their first visit will increase conversion rates and reduce the outflow to competitor sites.