User Behavioural Segmentation
The segments will be created by looking for relationships which exists between the users who are still using the application and the users who have uninstalled the application. Use these relationships to create user segments to identify retained users, and a segment of users who are still within the opportunity to be retained before they move to other alternatives.